requirements.txt
beautifulsoup4
aiohttp
openai
python-dotenv